Uniwersalnie sformatowana karta edycji w profilu. Każdy dymek ma cień i przycisk edycji danych znajdujących się w nim
