+++ BalansApp/app/src/main/java/com/example/firstcomposeap/ui/components/profile/ProfilTab.kt
@@ -1,6 +1,7 @@
 package com.example.firstcomposeap.ui.components.profile
 
 import android.R.attr.shadowColor
+import android.icu.util.UniversalTimeScale
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -33,6 +34,7 @@ import androidx.compose.ui.unit.sp
 import com.example.balansapp.ui.components.HeadText
 import com.example.balansapp.ui.service.LoginViewModel
 import com.example.balansapp.ui.service.data.Uzytkownik
+import com.example.firstcomposeap.ui.components.UniversalEditCard
 import java.nio.file.WatchEvent
 
 @Composable
@@ -44,80 +46,37 @@ fun ProfilTab (loginViewModel: LoginViewModel) {
     val user = loginViewModel.user
 
     UserInformationCard(user = user, onClick = {})
-    UserInformationCard(user = user, onClick = {})
-    UserInformationCard(user = user, onClick = {})
+
 }
+
 
 @Composable
 fun UserInformationCard(user: Uzytkownik?,
                         onClick: () -> Unit // kliknięcie wywoła edycję danych użytkownika
 ) {
-    var shadowColor = MaterialTheme.colorScheme.primary
+    UniversalEditCard(
+        data = {
+            Text("imie: ${user?.imie}", fontSize = 30.sp)
+            Text("nazwisko: ${user?.nazwisko}", fontSize = 30.sp)
 
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-            .drawBehind {
-                val shadowOffsetX = 8f
-                val shadowOffsetY = 8f
-                val shadowColor = shadowColor.copy(alpha = 0.25f)
-                drawRoundRect(
-                    color = shadowColor,
-                    topLeft = Offset(shadowOffsetX, shadowOffsetY),
-                    size = size,
-                    cornerRadius = CornerRadius(25f, 25f),
-                )
-            }
-            .shadow(
-                elevation = 10.dp,
-                shape = RoundedCornerShape(25.dp),
-                ambientColor = shadowColor.copy(alpha = 0.8f),
-                spotColor = shadowColor.copy(alpha = 0.8f)
-            )
-            .border(
-                width = 2.dp,
-                color = Color.Gray.copy(alpha = 0.4f),
-                shape = RoundedCornerShape(25.dp)
-            )
-            .background(Color.White, RoundedCornerShape(16.dp))
-            .padding(16.dp)
-    ) {
-        Row (modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically) {
+            Spacer(modifier = Modifier.height(5.dp))
+            Text("plec: ${user?.plec}", fontSize = 25.sp)
+            Text("wzrost: ${user?.wzrost}", fontSize = 25.sp)
+            Spacer(modifier = Modifier.height(5.dp))
 
-            Column (modifier = Modifier.weight(1f)) {
-                Text("imie: ${user?.imie}", fontSize = 30.sp)
-                Text("nazwisko: ${user?.nazwisko}", fontSize = 30.sp)
+            Text("email: ${user?.email}", fontSize = 25.sp)
 
-                Spacer(modifier = Modifier.height(5.dp))
-                Text("plec: ${user?.plec}", fontSize = 25.sp)
-                Text("wzrost: ${user?.wzrost}", fontSize = 25.sp)
-                Spacer(modifier = Modifier.height(5.dp))
-
-                Text("email: ${user?.email}", fontSize = 25.sp)
-                Text("id: ${user?.id}", fontSize = 30.sp)
+            Spacer(modifier = Modifier.height(25.dp))
+            Text("id: ${user?.id}", fontSize = 30.sp)
+            Text("waga: ${user?.waga}", fontSize = 30.sp)
+            Text("dania: ${user?.dania}", fontSize = 30.sp)
+            Text("aktualnyPlan: ${user?.aktualnyPlan}", fontSize = 30.sp)
+            Text("przyjaciele: ${user?.przyjaciele}", fontSize = 30.sp)
+        },
+        onClick = onClick
+    )
 
 
-                Text("waga: ${user?.waga}", fontSize = 30.sp)
-                Text("dania: ${user?.dania}", fontSize = 30.sp)
-                Text("aktualnyPlan: ${user?.aktualnyPlan}", fontSize = 30.sp)
-                Text("przyjaciele: ${user?.przyjaciele}", fontSize = 30.sp)
-            }
-            Column (horizontalAlignment = Alignment.End) {
-                FloatingActionButton(
-                    onClick = onClick,
-                    containerColor = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.size(42.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Edit,
-                        contentDescription = "Edytuj dane",
-                        tint = Color.White
-                    )
-                }
-            }
-        }
-    }
 }
+
+
